setupui/requirements.txt
attrs==21.3.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudflare==2.8.15
dnspython==2.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonlines==3.0.0
MarkupSafe==2.0.1
pip==21.1.2
PyYAML==6.0
requests==2.26.0
setuptools==57.0.0
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.2
wheel==0.36.2
